+++ src/main/java/shopping/domain/TokenProvider.java
@@ -1,0 +1,8 @@
+package shopping.domain;
+
+public interface TokenProvider {
+
+    String createToken(Long id);
+
+    Long decodeToken(String token);
+}
